AI Pathing (Implementing NavMeshes)
Spent too dang long trying to figure out why my AI isn't walking on the floor mesh that was made.
Turns out it was because the NavMesh Unity asset was needed, and the "baking" tool had to be JUST right otherwise it wouldn't work.
I now have a method of making floors that AI can walk and wander on. (May also have a method for implementing combat code)